Changed precision of calculations with adding bigDecimals

+++ src/main/java/com/currencyapp/app/controllers/RateRestController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
@@ -40,7 +41,7 @@ public class RateRestController {
             throw new MoreOrLessThanThreeLettersException();
         }
 
-        Map<String, Double> standardDeviationAndAverageMap = rateService.getStandardDeviationAndAverageMap(code,
+        Map<String, BigDecimal> standardDeviationAndAverageMap = rateService.getStandardDeviationAndAverageMap(code,
                 LocalDate.parse(startDate), LocalDate.parse(endDate));
         return new ModelAndView("resultView", "valuesToDisplay", standardDeviationAndAverageMap);
     }
